♻️ resilience,logging: Narrow bucket capacity to int
MemoryTokenBucket and RateLimitFilter both typed capacity as
PositiveFloat, but on MemoryTokenBucket the Pydantic annotation
was decorative (plain class, not a BaseModel) and the real guard
was a manual `if capacity <= 0` check.

Narrow capacity to int across both primitives. This matches the
industry standard for token-bucket burst size (Nginx `burst=`,
Envoy `max_tokens: uint32`, AWS API Gateway `burstLimit`,
Upstash `maxTokens`, Python `limits` / `pyrate_limiter`, and the
IETF RateLimit headers draft all use integer), and aligns the
standalone primitive with the existing `TokenBucket` algorithm
config which is already `PositiveInt`. Avoids the edge case
where `int(capacity)` of a fractional value truncates to 0 and
produces confusing "cost must be between 1 and 0" errors.

refill_rate stays `float` / `PositiveFloat`: rates are
continuous and must be expressible as sub-unit quantities
(e.g. 0.1/s for "1 per 10s").

Also tighten the MemoryTokenBucket invalid-config test to
pytest.raises(ValueError) and fix its stale docstring
(ValidationError -> ValueError).

+++ tests/logging/test_ratelimit.py
@@ -38,7 +38,7 @@ def _make_record(
 def test_defaults() -> None:
     """Test default configuration."""
     # Arrange
-    default_capacity = 5.0
+    default_capacity = 5
     default_refill = 1.0
     default_cost = 1.0
 
@@ -74,9 +74,7 @@ def test_config_property_is_frozen() -> None:
         (1, 1, -1),
     ],
 )
-def test_invalid_config(
-    capacity: float, refill_rate: float, cost: float
-) -> None:
+def test_invalid_config(capacity: int, refill_rate: float, cost: float) -> None:
     """Test non-positive values raise ValidationError."""
     # Act & Assert
     with pytest.raises(ValidationError, match="greater than"):
